updates next-auth to v5
