+++ frontend/apps/web/app/api/auth/[...nextauth]/route.ts
@@ -1,6 +1,1 @@
-import { getAuthOptions } from '@/api-only/auth-config';
-import NextAuth from 'next-auth';
-
-const handler = NextAuth(getAuthOptions());
-
-export { handler as GET, handler as POST };
+export { GET, POST } from './auth';
