+++ frontend/apps/web/app/layout.tsx
@@ -1,4 +1,3 @@
-import { getAuthOptions } from '@/api-only/auth-config';
 import '@/app/globals.css';
 import SiteFooter from '@/components/SiteFooter';
 import SiteHeader from '@/components/SiteHeader';
@@ -9,8 +8,8 @@ import { Toaster } from '@/components/ui/toaster';
 import { fontSans } from '@/libs/fonts';
 import { cn } from '@/libs/utils';
 import { Metadata } from 'next';
-import { getServerSession } from 'next-auth';
 import { ReactElement } from 'react';
+import { auth } from './api/auth/[...nextauth]/auth';
 
 export const metadata: Metadata = {
   title: 'Neosync',
@@ -23,7 +22,7 @@ export default async function RootLayout({
 }: {
   children: React.ReactNode;
 }): Promise<ReactElement> {
-  const session = await getServerSession(getAuthOptions());
+  const session = await auth();
   return (
     <html lang="en" suppressHydrationWarning>
       <head />
