+++ frontend/apps/web/package.json
@@ -42,11 +42,12 @@
     "cmdk": "^0.2.0",
     "cron-validate": "^1.4.5",
     "date-fns": "^2.30.0",
+    "date-fns-tz": "^2.0.0",
     "memoize-one": "^6.0.0",
     "nanoid": "^5.0.4",
     "neocron": "^0.4.8",
     "next": "^14.0.3",
-    "next-auth": "^4.24.4",
+    "next-auth": "^5.0.0-beta.4",
     "next-themes": "^0.2.1",
     "react": "18.2.0",
     "react-day-picker": "^8.9.1",
